test(step3): testFixture 분리
