+++ src/test/java/kitchenpos/ui/ProductRestControllerTest.java
@@ -3,7 +3,7 @@ package kitchenpos.ui;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import kitchenpos.application.ProductService;
-import kitchenpos.testfixture.TestFixture;
+import kitchenpos.testfixture.ProductFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,8 +39,8 @@ class ProductRestControllerTest {
     @Test
     @DisplayName("상품 생성 API")
     void create() throws Exception {
-        var request = TestFixture.createProduct("테스트", 1000L);
-        var response = TestFixture.copy(request);
+        var request = ProductFixture.createProduct("테스트", 1000L);
+        var response = ProductFixture.copy(request);
         given(productService.create(any())).willReturn(response);
 
         var result = mockMvc.perform(
@@ -61,7 +61,7 @@ class ProductRestControllerTest {
     @DisplayName("상품 가격 변경 API")
     void changePrice() throws Exception {
         var productId = UUID.randomUUID();
-        var response = TestFixture.createProduct(productId, "테스트", 1000L);
+        var response = ProductFixture.createProduct(productId, "테스트", 1000L);
         given(productService.changePrice(any(), any())).willReturn(response);
 
         var result = mockMvc.perform(
@@ -82,8 +82,8 @@ class ProductRestControllerTest {
     @DisplayName("상품 목록 조회 API")
     void findAll() throws Exception {
         var response = List.of(
-                TestFixture.createProduct("테스트1", 1000L),
-                TestFixture.createProduct("테스트2", 2000L)
+                ProductFixture.createProduct("테스트1", 1000L),
+                ProductFixture.createProduct("테스트2", 2000L)
         );
 
         given(productService.findAll()).willReturn(response);
